Add better guided fuzzer
